test/sql/requirements.txt
mysql-connector-python==2.0.2
wsgiref==0.1.2
